refactor(PlayerDao): Dao 합치기
